[eojin][update] directory 이름 수정 (example->sparta)
