--- src/main/java/com/example/dtogram/DtogramApplication.java
@@ -1,13 +1,0 @@
-package com.example.dtogram;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DtogramApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(DtogramApplication.class, args);
-    }
-
-}
