--- src/main/java/com/example/dtogram/common/security/UserDetailsServiceImpl.java
@@ -1,26 +1,0 @@
-package com.sparta.myblogbackend.security;
-
-import com.sparta.myblogbackend.entity.User;
-import com.sparta.myblogbackend.repository.UserRepository;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserDetailsServiceImpl implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    public UserDetailsServiceImpl(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        User user = userRepository.findByUsername(username)
-                .orElseThrow(() -> new UsernameNotFoundException("Not Found " + username));
-
-        return new UserDetailsImpl(user);
-    }
-}
