--- src/main/java/com/example/dtogram/post/dto/PostRequestDto.java
@@ -1,11 +1,0 @@
-package com.example.dtogram.post.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class PostRequestDto {
-    private String title;
-    private String content;
-}
