--- src/main/java/com/example/dtogram/post/repository/PostRepository.java
@@ -1,12 +1,0 @@
-package com.example.dtogram.post.repository;
-import com.example.dtogram.post.entity.Post;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface PostRepository extends JpaRepository<Post, Long> {
-    List<Post> findAllByOrderByModifiedAtDesc();
-    List<Post> findAllByContentsContainingOrderByModifiedAtDesc(String keyword);
-}
