--- src/main/java/com/example/dtogram/reply/controller/CommentController.java
@@ -1,51 +1,0 @@
-package com.example.dtogram.Reply.controller;
-
-import com.sparta.myblogbackend.dto.CommentRequestDto;
-import com.sparta.myblogbackend.dto.CommentResponseDto;
-import com.sparta.myblogbackend.entity.Blog;
-import com.sparta.myblogbackend.security.UserDetailsImpl;
-import com.sparta.myblogbackend.service.CommentService;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@Slf4j
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api")
-public class CommentController {
-
-    private final CommentService commentService;
-
-    @PostMapping("/comments")
-    public CommentResponseDto createComment(@RequestParam Long blog_id, @RequestBody CommentRequestDto requestDto, @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        return commentService.createComment(requestDto, userDetails.getUser(), blog_id);
-    }
-
-    @PutMapping("/comment")
-    @ResponseBody
-    public CommentResponseDto updateComment(@RequestParam Long id, @RequestBody CommentRequestDto requestDto, @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        return commentService.updateComment(id, requestDto, userDetails.getUser());
-    }
-
-    @DeleteMapping("/comment")
-    public ResponseEntity<String> deleteComment(@RequestParam Long id, @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        commentService.deleteComment(id, userDetails.getUser());
-        return ResponseEntity.status(HttpStatus.OK).body("댓글 삭제 성공");
-    }
-
-    @PostMapping("/comment/like")
-    public void like(@RequestParam Long id, @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        commentService.like(id, userDetails.getUser().getId());
-    }
-
-    @GetMapping("/comment/like")
-    public boolean isLiked(@RequestParam Long id, @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        return commentService.isLiked(id, userDetails.getUser().getId());
-    }
-}
