--- src/main/java/com/example/dtogram/reply/dto/CommentRequestDto.java
@@ -1,8 +1,0 @@
-package com.example.dtogram.Reply.dto;
-
-import lombok.Getter;
-
-@Getter
-public class CommentRequestDto {
-    private String comments;
-}
