--- src/main/java/com/example/dtogram/user/dto/LoginRequestDto.java
@@ -1,11 +1,0 @@
-package com.example.dtogram.User.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Setter
-@Getter
-public class LoginRequestDto {
-    private String username;
-    private String password;
-}
