--- src/main/java/com/example/dtogram/user/dto/UserInfoDto.java
@@ -1,11 +1,0 @@
-package com.example.dtogram.User.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class UserInfoDto {
-    String username;
-    boolean isAdmin;
-}
