--- src/main/java/com/example/dtogram/user/entity/User.java
@@ -1,53 +1,0 @@
-package com.example.dtogram.user.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@Table(name = "users")
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, unique = true)
-    private String username;
-
-    @Column(nullable = false)
-    private String password;
-
-    @Column(nullable = false, unique = true)
-    private String email;
-
-    @Column(nullable = false)
-    @Enumerated(value = EnumType.STRING)
-    private com.example.dtogram.User.entity.UserRoleEnum role;
-
-    @OneToMany(fetch = FetchType.LAZY, mappedBy = "user", cascade = CascadeType.REMOVE)
-    private List<BlogLike> blogLikeList = new ArrayList<>();
-
-    @OneToMany(fetch = FetchType.LAZY, mappedBy = "user", cascade = CascadeType.REMOVE)
-    private List<CommentLike> commentLikeList = new ArrayList<>();
-
-    public User(String username, String password, String email, com.example.dtogram.User.entity.UserRoleEnum role) {
-        this.username = username;
-        this.password = password;
-        this.email = email;
-        this.role = role;
-    }
-
-    public void mappingBlogLike(BlogLike blogLike) { // 유저가 해당 좋아요를 눌렀는지 확인
-        this.blogLikeList.add(blogLike);
-    }
-    public void mappingCommentLike(CommentLike commentLike) { // 유저가 해당 좋아요를 눌렀는지 확인
-        this.commentLikeList.add(commentLike);
-    }
-}
