+++ src/main/java/com/sparta/dtogram/DtogramApplication.java
@@ -1,0 +1,13 @@
+package com.sparta.dtogram;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DtogramApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(DtogramApplication.class, args);
+    }
+
+}
