+++ src/main/java/com/sparta/dtogram/like/entity/PostLike.java
@@ -1,0 +1,24 @@
+package com.sparta.dtogram.like.entity;
+
+import com.sparta.dtogram.user.entity.User;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+@Table(name = "postLike") // 매핑할 테이블의 이름을 지정
+@NoArgsConstructor
+public class PostLike {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+
+
+}
