+++ src/main/java/com/sparta/dtogram/post/dto/PostRequestDto.java
@@ -1,0 +1,11 @@
+package com.sparta.dtogram.post.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class PostRequestDto {
+    private String title;
+    private String content;
+}
