+++ src/main/java/com/sparta/dtogram/reply/dto/CommentRequestDto.java
@@ -1,0 +1,8 @@
+package com.sparta.dtogram.reply.dto;
+
+import lombok.Getter;
+
+@Getter
+public class CommentRequestDto {
+    private String comments;
+}
