+++ src/main/java/com/sparta/dtogram/reply/repository/CommentRepository.java
@@ -1,0 +1,9 @@
+package com.sparta.dtogram.reply.repository;
+
+import com.sparta.myblogbackend.entity.Comment;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+}
