+++ src/main/java/com/sparta/dtogram/user/dto/LoginRequestDto.java
@@ -1,0 +1,11 @@
+package com.sparta.dtogram.user.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+public class LoginRequestDto {
+    private String username;
+    private String password;
+}
