+++ src/main/java/com/sparta/dtogram/user/dto/SignupRequestDto.java
@@ -1,0 +1,23 @@
+package com.sparta.dtogram.user.dto;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Pattern;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class SignupRequestDto {
+    @NotBlank
+    @Pattern(regexp = "[a-z0-9]{4,10}")
+    private String username;
+    @NotBlank
+    @Pattern(regexp = "[a-zA-Z0-9!@#$%^&*]{8,15}")
+    private String password;
+    @Email
+    @NotBlank
+    private String email;
+    private boolean admin = false;
+    private String adminToken = "";
+}
