+++ src/main/java/com/sparta/dtogram/user/dto/UserInfoDto.java
@@ -1,0 +1,11 @@
+package com.sparta.dtogram.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class UserInfoDto {
+    String username;
+    boolean isAdmin;
+}
