+++ src/main/java/com/sparta/dtogram/user/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package com.sparta.dtogram.user.repository;
+
+import com.sparta.myblogbackend.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByUsername(String username);
+
+    Optional<User> findByEmail(String email);
+}
