--- src/test/java/com/example/dtogram/DtogramApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.dtogram;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DtogramApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
