+++ src/test/java/com/sparta/dtogram/DtogramApplicationTests.java
@@ -1,0 +1,13 @@
+package com.sparta.dtogram;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DtogramApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
